update:  change color theme on bottom bar and header, add confirm function, confirm logout added

+++ stores/useAuth.ts
@@ -1,5 +1,7 @@
 import { create } from "zustand";
 import useSecureStore from "../utils/useSecureStore";
+import { Toast } from "toastify-react-native";
+import { confirm } from "../utils/helpers";
 
 type AuthState = {
     isAuthenticated: boolean;
@@ -27,11 +29,15 @@ export const useAuth = create<AuthState>((set) => ({
         set({ isAuthenticated: true, user: user, token: token });
         await useSecureStore().save("user", JSON.stringify(user));
         await useSecureStore().save("token", token);
+        Toast.success('Login telah berhasil!')
     },
     logout: async () => {
+        const confirmation = await confirm('Yakin ingin keluar akun?', 'Anda akan keluar dari akun user.', 'Keluar', 'destructive');
+        if (!confirmation) return;
         set({ isAuthenticated: false, user: null, token: null });
         await useSecureStore().deleteItem("token");
         await useSecureStore().deleteItem("user");
+        Toast.info('Berhasil logout dari akun')
     },
     persistLogin: async () => {
         const storedUser = await useSecureStore().getValue("user").then(data => JSON.parse(data));
